fix xa breakage on 64bit Linux
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@82087 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/dfsound/psemuxa.h
@@ -8,7 +8,7 @@
 
 typedef struct
 {
-	long	y0, y1;
+	int	y0, y1;
 } ADPCM_Decode_t;
 
 typedef struct
